Serve the highlight row from one shared projection
The book-details page and the in-book highlight search rendered the same
highlight through the same `Highlight` response schema, but each owned a
private copy of the whole path: `HighlightView` and `SearchHighlightView`
declared the same 13 fields, both `_highlight_view` assemblers mapped them
identically, and both `_build_highlight_schema` builders were byte-identical
apart from their parameter annotation. A field added to the row had to be
written out in six places to reach the wire.

Collapse them onto one projection: `HighlightRow` plus the `TagRef`,
`FlashcardRef` and `BookmarkView` shapes several views embed, with one
ORM-to-DTO mapper and one DTO-to-schema builder.

The shared DTOs sit in `application/common/queries/`, which still matches the
`queries-are-dead-ends` wildcard, so commands remain unable to import them.
Label resolution keeps going through `LabelResolutionService` rather than
moving into SQL (ADR-0001, rule 1).

No API contract change: the generated OpenAPI document is byte-identical.
