+++ backend/src/application/reading/queries/reading_sessions.py
@@ -8,16 +8,8 @@ from dataclasses import dataclass
 from datetime import datetime as dt
 from typing import Protocol
 
+from src.application.common.queries.highlight_row import HighlightLabelView
 from src.domain.common.value_objects.ids import BookId, UserId
-
-
-@dataclass(frozen=True)
-class HighlightLabelView:
-    """A highlight's effective label, resolved by the domain rather than by SQL."""
-
-    highlight_style_id: int
-    text: str | None
-    ui_color: str | None
 
 
 @dataclass(frozen=True)
