+++ backend/src/infrastructure/reading/routers/highlights.py
@@ -13,7 +13,6 @@ from src.application.reading.commands.highlights.highlight_upload_use_case impor
 )
 from src.application.reading.queries.highlight_search import (
     SearchChapterView,
-    SearchHighlightView,
 )
 from src.application.reading.queries.highlight_search_use_case import (
     HighlightSearchUseCase,
@@ -22,60 +21,17 @@ from src.core import container
 from src.domain.identity.entities.user import User
 from src.infrastructure.common.di import inject_use_case
 from src.infrastructure.identity.dependencies import get_current_user
-from src.infrastructure.learning.schemas import Flashcard
 from src.infrastructure.reading.schemas import (
     BookHighlightSearchResponse,
     ChapterWithHighlights,
-    Highlight,
     HighlightDeleteRequest,
     HighlightDeleteResponse,
-    HighlightLabel,
     HighlightUploadRequest,
     HighlightUploadResponse,
 )
-from src.infrastructure.tagging.schemas import TagInBook
+from src.infrastructure.reading.schemas.highlight_builders import build_highlight_schema
 
 router = APIRouter(prefix="", tags=["highlights"])
-
-
-def _build_highlight_schema(highlight: SearchHighlightView) -> Highlight:
-    """Build the Highlight schema from a match in the search read model."""
-    return Highlight(
-        id=highlight.id,
-        book_id=highlight.book_id,
-        chapter_id=highlight.chapter_id,
-        text=highlight.text,
-        chapter=highlight.chapter_name,
-        chapter_number=highlight.chapter_number,
-        page=highlight.page,
-        datetime=highlight.datetime,
-        label=HighlightLabel(
-            highlight_style_id=highlight.label.highlight_style_id,
-            text=highlight.label.text,
-            ui_color=highlight.label.ui_color,
-        )
-        if highlight.label
-        else None,
-        flashcards=[
-            Flashcard(
-                id=card.id,
-                user_id=card.user_id,
-                book_id=card.book_id,
-                highlight_id=card.highlight_id,
-                chapter_id=card.chapter_id,
-                note_id=card.note_id,
-                question=card.question,
-                answer=card.answer,
-            )
-            for card in highlight.flashcards
-        ],
-        tags=[
-            TagInBook(id=tag.id, name=tag.name, tag_group_id=tag.tag_group_id)
-            for tag in highlight.tags
-        ],
-        created_at=highlight.created_at,
-        updated_at=highlight.updated_at,
-    )
 
 
 def _build_chapter_schema(chapter: SearchChapterView) -> ChapterWithHighlights:
@@ -89,7 +45,7 @@ def _build_chapter_schema(chapter: SearchChapterView) -> ChapterWithHighlights:
         chapter_number=chapter.chapter_number,
         parent_id=None,
         start_position=None,
-        highlights=[_build_highlight_schema(highlight) for highlight in chapter.highlights],
+        highlights=[build_highlight_schema(highlight) for highlight in chapter.highlights],
         created_at=chapter.created_at,
         updated_at=chapter.updated_at,
     )
